#4 and #7 - refactored
